Add files via upload
1.Welcome Page
It consists of button and frame. We had used images also. It is the starting page. We had linked all other pages with this and import all those pages. It consists of About button , Login button, New User, Exit button (which are detailed below) .
In this firstly We had created its gui . We had applied the concept of error handling and proper use of widgets has been applied in it.

2. About us
It consists of simple images, emojis and text. It just specify that this project has been made by us. Proper use of gui has been implemented.

3. Login
It consists of simple images, buttons and frames. In this concept of gui , database and error handling has been used. If all the details are filled correctly then only user can proceed further.

4. New user
It consists of all the widgets, images. It stores the data provided by the user. In this proper use of gui, error handling and database is there. We had linked this with sqlite and other pages of the project.

5. Main
It consists of simple widgets like frame, buttons, label, text, entry, radio buttons, etc. It is linked with only Login page. If the user enter correct details and it is found in database then only this page will open. In this proper use of gui and error handling is used.